+++ lib/pg_canary/rules/definitions/array_search_without_gin.rb
@@ -11,10 +11,6 @@ module PgCanary
         true
       end
 
-      def size_dependent?
-        false
-      end
-
       ARRAY_OPS = %w[@> <@ &&].freeze
 
       def check(query)