+++ lib/pg_canary/rules/definitions/huge_in_list.rb
@@ -12,10 +12,6 @@ module PgCanary
         true
       end
 
-      def size_dependent?
-        false
-      end
-
       def self.options
         { threshold: 500 }
       end