+++ lib/pg_canary/rules/definitions/or_across_columns.rb
@@ -10,10 +10,6 @@ module PgCanary
         false
       end
 
-      def size_dependent?
-        true
-      end
-
       def check(query)
         detections = []
         query.each_scope do |scope|