+++ lib/pg_canary/rules/definitions/order_by_random.rb
@@ -9,10 +9,6 @@ module PgCanary
         true
       end
 
-      def size_dependent?
-        false
-      end
-
       def check(query)
         detections = []
         query.each_scope do |scope|