+++ lib/pg_canary/rules/definitions/unindexed_join.rb
@@ -12,10 +12,6 @@ module PgCanary
         false
       end
 
-      def size_dependent?
-        true
-      end
-
       def check(query)
         detections = []
         query.each_scope do |scope|